backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
cx-Oracle==8.3.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
PyJWT==2.8.1
bcrypt==4.1.1
python-jose==3.3.0
email-validator==2.1.0
fastapi-cors==0.0.6
python-dateutil==2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
